@@ -20,9 +20,9 @@
     "postinstall": "npx electron-builder install-app-deps",
     "ng": "ng",
     "start": "node hooks/environments/set_profile.js && npm-run-all -p ng:serve electron:serve",
-    "build": "node hooks/environments/set_profile.js && ng build && npm run electron:tsc",
-    "build:prod": "node hooks/environments/set_profile.js && ng build --prod && npm run electron:tsc",
-    "ng:serve": "ng serve -o",
+    "build": "node hooks/environments/set_profile.js && webpack && npm run electron:tsc",
+    "build:prod": "node hooks/environments/set_profile.js && webpack --prod && npm run electron:tsc",
+    "ng:serve": "webpack-dev-server --port=4200",
     "electron:tsc": "tsc main.ts",
     "electron:serve": "wait-on http-get://localhost:4200/ && npm run electron:tsc && electron . --serve",
     "electron:local": "npm run build:prod && electron .",
@@ -33,8 +33,14 @@
     "pree2e:build": "webdriver-manager update --gecko false && npm run ng serve",
     "pree2e:protractor": "wait-on http-get://localhost:4200/ && protractor ./protractor.conf.js",
     "e2e": "npm-run-all -p -r pree2e:build pree2e:protractor"
+
   },
-  "dependencies": {},
+  "dependencies": {
+    "@types/johnny-five": "0.0.36",
+    "electron": "^2.0.0",
+    "johnny-five": "^0.15.0",
+    "serialport": "^6.2.0"
+  },
   "devDependencies": {
     "@angular/cli": "1.7.4",
     "@angular/common": "5.2.9",
@@ -57,8 +63,8 @@
     "core-js": "2.5.5",
     "cross-env": "5.1.4",
     "dotenv": "5.0.1",
-    "electron": "1.8.4",
     "electron-builder": "20.8.1",
+    "electron-rebuild": "^1.7.3",
     "electron-reload": "1.2.2",
     "fs-extra": "5.0.0",
     "jasmine-core": "2.99.1",
@@ -78,6 +84,26 @@
     "typescript": "2.6.2",
     "wait-on": "2.1.0",
     "webdriver-manager": "12.0.6",
-    "zone.js": "0.8.26"
+    "zone.js": "0.8.26",
+    "webpack-dev-server": "~2.11.0",
+    "@angular-devkit/core": "0.3.2",
+    "@ngtools/webpack": "1.10.2",
+    "webpack": "~3.11.0",
+    "autoprefixer": "^7.2.3",
+    "file-loader": "^1.1.5",
+    "html-webpack-plugin": "^2.29.0",
+    "less-loader": "^4.0.5",
+    "postcss-import": "^11.0.0",
+    "postcss-loader": "^2.0.10",
+    "postcss-url": "^7.1.2",
+    "raw-loader": "^0.5.1",
+    "sass-loader": "^6.0.6",
+    "istanbul-instrumenter-loader": "^3.0.0",
+    "style-loader": "^0.19.1",
+    "stylus-loader": "^3.0.1",
+    "url-loader": "^0.6.2",
+    "circular-dependency-plugin": "^4.2.1",
+    "copy-webpack-plugin": "~4.4.1",
+    "uglifyjs-webpack-plugin": "^1.1.8"
   }
 }
